fix wrong param to grow array in line

+++ chunk.c
@@ -19,7 +19,7 @@ void write_chunk(Chunk *c, uint8_t byte,int line)
         int old_capacity = c->capacity;
         c->capacity = GROW_CAPACITY(old_capacity);
         c->code = GROW_ARRAY(uint8_t,c->code,old_capacity,c->capacity);
-        c->line = GROW_ARRAY(uint8_t,c->line,old_capacity,c->capacity);
+        c->line = GROW_ARRAY(int,c->line,old_capacity,c->capacity);
     }
     
     c->code[c->count] = byte;
